+++ Minter/src/test/java/com/hida/service/MinterServiceImplTest.java
@@ -1,8 +1,8 @@
 package com.hida.service;
 
-import com.hida.dao.DefaultSettingRepository;
-import com.hida.dao.PidRepository;
-import com.hida.dao.UsedSettingRepository;
+import com.hida.repositories.DefaultSettingRepository;
+import com.hida.repositories.PidRepository;
+import com.hida.repositories.UsedSettingRepository;
 import com.hida.model.DefaultSetting;
 import com.hida.model.NotEnoughPermutationsException;
 import com.hida.model.Pid;
